==> 1. Preprocessing.ipynb <==
#import library
import numpy as np
import pandas as pd
import pandas_profiling
from pandas_profiling import ProfileReport

#import dan menampilkan isi data.csv
df = pd.read_csv("csv/data.csv")
df

#eksplorasi Data Menggunakan Pandas Profiling Report
pandas_report = df.profile_report(sort='None', html={'style':{'full_width': True}})
pandas_report

#melakukan drop pada beberapa kolom yang tidak digunakan
df = df.drop(['SEMESTER','KEHADIRAN','COMPONENTNAME','POINT','TOTALPOINT','PLONAME','PLONUMBER','COURSEID'], axis = 1)

#menampilkan info apakah terdapat data null pada dataset
df.isna().sum()

#melakukan drop pada data yg memiliki null
df.dropna(inplace=True)

#menghitung banyaknya data yang duplicate
df.duplicated().sum()

#menyimpan data pertama dan melakukan drop pada sisa data yang sama (duplicate)
df = df.drop_duplicates(subset=None, keep='first')

#mereset indeks pada tabel & menyimpan data yang sudah bersih ke dalam data_clean.csv
df = df.reset_index(drop=True)
df.to_csv('csv/data_clean.csv', index = False)
df

#menampilkan Data Menggunakan Pandas Profiling Report
pandas_report = df.profile_report(sort='None', html={'style':{'full_width': True}})
pandas_report

#membagi dan menyimpan data nilai mahasiswa berdasarkan angkatannya
df_1617 = df[df['ANGKATAN'] == 1617]
df_1617.to_csv('csv/data_1617.csv', index = False)

df_1718 = df[df['ANGKATAN'] == 1718]
df_1718.to_csv('csv/data_1718.csv', index = False)

df_1819 = df[df['ANGKATAN'] == 1819]
df_1819.to_csv('csv/data_1819.csv', index = False)
