==> 2. Preprocessing.ipynb <==
# import library
import numpy as np
import pandas as pd
import pandas_profiling
from pandas_profiling import ProfileReport

import warnings
warnings.simplefilter(action='ignore', category=pd.errors.PerformanceWarning)

#function mengubah nama mata kuliah yang ekivalensi dari kurikulum 2016 ke 2020
def ubahNamaMK(df):
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'LOGIKA MATEMATIKA A', value = 'LOGIKA MATEMATIKA')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'KALKULUS IB', value = 'KALKULUS')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'LITERASI TIK', value = 'STATISTIKA')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'MATEMATIKA DISKRIT A', value = 'MATEMATIKA DISKRIT')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'DASAR ALGORITMA DAN PEMROGRAMAN', value = 'ALGORITMA PEMROGRAMAN')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'KALKULUS IIB', value = 'KALKULUS LANJUT')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'DESAIN DAN ANALISIS ALGORITMA', value = 'ANALISIS KOMPLEKSITAS ALGORITMA')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'ANALISIS DAN PERANCANGAN PERANGKAT LUNAK', value = 'RPL: ANALISIS KEBUTUHAN')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'PROBABILITAS DAN STATISTIKA A', value = 'TEORI PELUANG')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'KECERDASAN BUATAN', value = 'PENGANTAR KECERDASAN BUATAN')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'SISTEM OPERASI A', value = 'SISTEM OPERASI')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'TEORI BAHASA & AUTOMATA', value = 'TEORI BAHASA DAN AUTOMATA')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'INTERAKSI MANUSIA DAN KOMPUTER', value = 'INTERAKSI MANUSIA KOMPUTER')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'INTERAKSI MANUSIA DAN KOMPUTER A', value = 'INTERAKSI MANUSIA KOMPUTER')
    df['MATAKULIAH'] = df['MATAKULIAH'].replace(to_replace = 'PEMROGRAMAN BERORIENTASI OBYEK A', value = 'PEMROGRAMAN BERORIENTASI OBJEK')
    
#function membuat dataset baru dengan nama mata kuliah sebagai kolomnya
def datasetBaru(df):
    #menentukan nama file (pada folder csv) berdasarkan angkatan
    if df['ANGKATAN'][0] == 1617:
        file1 = 'csv/list_mk_angkatan16.csv'
        file2 = 'csv/data_baru1617.csv'
    elif df['ANGKATAN'][0] == 1718:
        file1 = 'csv/list_mk_angkatan17.csv'
        file2 = 'csv/data_baru1718.csv'
    elif df['ANGKATAN'][0] == 1819:
        file1 = 'csv/list_mk_angkatan18.csv'
        file2 = 'csv/data_baru1819.csv'
    
    #melakukan drop pada beberapa kolom
    df = df.drop(['NIM','ANGKATAN','INDEXPOINT'], axis = 1)
    #menghapus data yang duplicate
    df = df.drop_duplicates('MATAKULIAH', keep = 'first')
    df = df.reset_index(drop=True)
    
    #menyimpan file yg berisi KODE_MK dan MATAKULIAH
    df.to_csv(file1, index = False)
    
    #mengubah kolom KODE_MK menjadi list
    list_mk = df['MATAKULIAH'].tolist()
    df_fix = pd.DataFrame()

    #membuat kolom pada dataset baru (df_fix)
    df_fix['NIM'] = np.nan
    df_fix[list_mk] = np.nan
    df_fix.to_csv(file2, index = False)

#function mengubah index nilai alfabet menjadi numerik
def encoderNilai(df):
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'A', value = 4)
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'AB', value = 3.5)
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'B', value = 3)
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'BC', value = 2.5)
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'C', value = 2)
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'D', value = 1)
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'E', value = 0)
    df['INDEXPOINT'] = df['INDEXPOINT'].replace(to_replace = 'T', value = 0)

#function pemindahan dari dataset lama ke dataset baru
def pindahDataseBaru(df,df_fix):
    #membuat baris ke-0 pada dataframe_fix
    df_fix.loc[0] = np.nan
    df_fix['NIM'][0] = df['NIM'][0]
    
    i = 0
    #melakukan perulangan untuk memasukkan nilai dari dataset lama ke dataset baru
    for index, row in df.iterrows():
        if (df_fix['NIM'] == df['NIM'][index]).any() == True:
            #perulangan untuk memasukkan nilai sesuai dengan kolom pada df_fix
            for column in df_fix:
                if df['MATAKULIAH'][index] == column:
                    #kondisi jika nilai MK yg akan dimasukkan tidak kosong (mengulang mata kuliah)
                    if np.isnan(df_fix[column][i]) == True:
                        nilai = df.loc[index, 'INDEXPOINT']
                        df_fix[column][i] = nilai
                    elif np.isnan(df_fix[column][i]) == False:
                        nilai1 = df.loc[index, 'INDEXPOINT']
                        nilai2 = df_fix[column][i]
                        if nilai2 < nilai1:
                            df_fix[column][i] = nilai1          
        else:
            i = i + 1
            df_fix.loc[i] = np.nan
            df_fix['NIM'][i] = df['NIM'][index]
            #perulangan untuk memasukkan nilai sesuai dengan kolom pada df_fix
            for column in df_fix:
                if df['MATAKULIAH'][index] == column:
                    nilai = df.loc[index, 'INDEXPOINT']
                    df_fix[column][i] = nilai
    return df_fix

#melakukan import isi data_1617.csv dan membuat file list mata kuliah pada angkatan 2016
df = pd.read_csv("csv/data_1617.csv")
ubahNamaMK(df)
datasetBaru(df)

#melakukan import isi data_baru1617.csv, melakukan encoderNilai dan pemindahan data ke dataset baru
df_fix = pd.read_csv("csv/data_baru1617.csv")
encoderNilai(df)
df_fix = pindahDataseBaru(df,df_fix)

#mengisi data null menjadi 0 (mahasiswa yg tidak mengambil suatu mkpil atau belum mengambil)
df_fix = df_fix.fillna(0)
df_fix.to_csv('csv/data_fix1617.csv', index = False)
df_fix

#melakukan import isi data_1718.csv, mengubah ubahNamaMK dan membuat datasetBaru
df = pd.read_csv("csv/data_1718.csv")
ubahNamaMK(df)
datasetBaru(df)

#melakukan import isi data_baru1718.csv, melakukan encoderNilai dan pemindahan data ke dataset baru
df_fix = pd.read_csv("csv/data_baru1718.csv")
encoderNilai(df)
df_fix = pindahDataseBaru(df,df_fix)

#mengisi data null menjadi 0 (mahasiswa yg tidak mengambil suatu mkpil atau belum mengambil)
df_fix = df_fix.fillna(0)
df_fix.to_csv('csv/data_fix1718.csv', index = False)
df_fix

#melakukan import isi data_1819.csv, mengubah ubahNamaMK dan membuat datasetBaru
df = pd.read_csv("csv/data_1819.csv")
ubahNamaMK(df)
datasetBaru(df)

#melakukan import isi data_baru1819.csv, melakukan encoderNilai dan pemindahan data ke dataset baru
df_fix = pd.read_csv("csv/data_baru1819.csv")
encoderNilai(df)
df_fix = pindahDataseBaru(df,df_fix)

#mengisi data null menjadi 0 (mahasiswa yg tidak mengambil suatu mkpil atau belum mengambil)
df_fix = df_fix.fillna(0)
df_fix.to_csv('csv/data_fix1819.csv', index = False)
df_fix
