==> 3. Preprocessing Data Training.ipynb <==
# import library
import numpy as np
import pandas as pd

#function memilih data nilai mahasiswa yg sudah lulus mk Tugas Akhir
def dataMhsLulus(df1,df2):
    #mengambil data nilai mahasiswa yang sudah lulus Tugas Akhir dengan indeksnilai min C
    df_lulus1 = df1[df1['TUGAS AKHIR'] != 0]
    df_lulus1 = df_lulus1[df_lulus1['TUGAS AKHIR'] != 1]
    
    df_lulus2 = df2[df2['TUGAS AKHIR'] != 0]
    df_lulus2 = df_lulus2[df_lulus2['TUGAS AKHIR'] != 1]
    
    #menggabungkan dataset mahasiswa 2016 dan 2017
    df_lulus = pd.concat([df_lulus1,df_lulus2])
    df_lulus = df_lulus.reset_index(drop=True)
    
    return df_lulus

#import data dan memanggil function dataMhsLulus
df1 = pd.read_csv("csv/data_fix1617.csv")
df2 = pd.read_csv("csv/data_fix1718.csv")
df_lulus = dataMhsLulus(df1,df2)
df_lulus

#mengambil data dari beberapa kolom mk wajib Intelligent System
df_wajib_is = df_lulus.filter(['LOGIKA MATEMATIKA',
                               'MATEMATIKA DISKRIT',
                               'ALGORITMA PEMROGRAMAN',
                               'ANALISIS KOMPLEKSITAS ALGORITMA',
                               'STRUKTUR DATA',
                               'PENGANTAR KECERDASAN BUATAN',
                               'TEORI BAHASA DAN AUTOMATA',
                               'PEMBELAJARAN MESIN',
                               'PEMROGRAMAN BERORIENTASI OBJEK'])
df_wajib_is

#mengambil data dari beberapa kolom mk wajib Cyber Physical System
df_wajib_cps = df_lulus.filter(['SISTEM DIGITAL',
                                'ORGANISASI DAN ARSITEKTUR KOMPUTER',
                                'SISTEM OPERASI',
                                'JARINGAN KOMPUTER',
                                'SISTEM PARALEL DAN TERDISTRIBUSI'])
df_wajib_cps

#mengambil data dari beberapa kolom mk wajib Software Engineering
df_wajib_se = df_lulus.filter(['PEMODELAN BASIS DATA',
                               'RPL: ANALISIS KEBUTUHAN',
                               'SISTEM BASIS DATA',
                               'INTERAKSI MANUSIA KOMPUTER'])
df_wajib_se

#mengambil data dari beberapa kolom mk wajib Data Science
df_wajib_ds = df_lulus.filter(['KALKULUS',
                               'STATISTIKA',
                               'KALKULUS LANJUT',
                               'MATRIKS DAN RUANG VEKTOR',
                               'TEORI PELUANG'])
df_wajib_ds

#membuat dataset baru yg hanya berisikan kolom mata kuliah dasar, wajib icm, wajib side, dan wajib tele
df = df_lulus
df_lulus = pd.concat([df_wajib_is, df_wajib_cps, df_wajib_se, df_wajib_ds], axis=1)

#setiap mk wajib IS, CPS, SE dan DS dikali bobot sksnya masing - masing
df_wajib_is['LOGIKA MATEMATIKA'] = df_wajib_is['LOGIKA MATEMATIKA'] * 3
df_wajib_is['MATEMATIKA DISKRIT'] = df_wajib_is['MATEMATIKA DISKRIT'] * 3
df_wajib_is['ALGORITMA PEMROGRAMAN'] = df_wajib_is['ALGORITMA PEMROGRAMAN'] * 4
df_wajib_is['ANALISIS KOMPLEKSITAS ALGORITMA'] = df_wajib_is['ANALISIS KOMPLEKSITAS ALGORITMA'] * 2
df_wajib_is['STRUKTUR DATA'] = df_wajib_is['STRUKTUR DATA'] * 4
df_wajib_is['PENGANTAR KECERDASAN BUATAN'] = df_wajib_is['PENGANTAR KECERDASAN BUATAN'] * 3
df_wajib_is['TEORI BAHASA DAN AUTOMATA'] = df_wajib_is['TEORI BAHASA DAN AUTOMATA'] * 3
df_wajib_is['PEMBELAJARAN MESIN'] = df_wajib_is['PEMBELAJARAN MESIN'] * 3
df_wajib_is['PEMROGRAMAN BERORIENTASI OBJEK'] = df_wajib_is['PEMROGRAMAN BERORIENTASI OBJEK'] * 4

df_wajib_cps['SISTEM DIGITAL'] = df_wajib_cps['SISTEM DIGITAL'] * 3
df_wajib_cps['ORGANISASI DAN ARSITEKTUR KOMPUTER'] = df_wajib_cps['ORGANISASI DAN ARSITEKTUR KOMPUTER'] * 3
df_wajib_cps['SISTEM OPERASI'] = df_wajib_cps['SISTEM OPERASI'] * 3
df_wajib_cps['JARINGAN KOMPUTER'] = df_wajib_cps['JARINGAN KOMPUTER'] * 4
df_wajib_cps['SISTEM PARALEL DAN TERDISTRIBUSI'] = df_wajib_cps['SISTEM PARALEL DAN TERDISTRIBUSI'] * 4

df_wajib_se['PEMODELAN BASIS DATA'] = df_wajib_se['PEMODELAN BASIS DATA'] * 3
df_wajib_se['RPL: ANALISIS KEBUTUHAN'] = df_wajib_se['RPL: ANALISIS KEBUTUHAN'] * 2
df_wajib_se['SISTEM BASIS DATA'] = df_wajib_se['SISTEM BASIS DATA'] * 3
df_wajib_se['INTERAKSI MANUSIA KOMPUTER'] = df_wajib_se['INTERAKSI MANUSIA KOMPUTER'] * 3

df_wajib_ds['KALKULUS'] = df_wajib_ds['KALKULUS'] * 3
df_wajib_ds['STATISTIKA'] = df_wajib_ds['STATISTIKA'] * 2
df_wajib_ds['KALKULUS LANJUT'] = df_wajib_ds['KALKULUS LANJUT'] * 3
df_wajib_ds['MATRIKS DAN RUANG VEKTOR'] = df_wajib_ds['MATRIKS DAN RUANG VEKTOR'] * 3
df_wajib_ds['TEORI PELUANG'] = df_wajib_ds['TEORI PELUANG'] * 3

#penghitungan nilai akhir setiap KK (((MK * SKS) + MK lainnya pada kk yang sama) / total bobot sks)
nilai_tiap_kk = pd.DataFrame(columns = ['IS', 'CPS', 'SE', 'DS'])

for index, row in df_wajib_is.iterrows():
    hasil = 0
    nilai_tiap_kk.loc[index] = np.nan
    for column in df_wajib_is:
        hasil_convert = df_wajib_is.loc[index, column]
        hasil = hasil + hasil_convert
    hasil = hasil / 29
    nilai_tiap_kk['IS'][index] = hasil
    
for index, row in df_wajib_cps.iterrows():
    hasil = 0
    for column in df_wajib_cps:
        hasil_convert = df_wajib_cps.loc[index, column]
        hasil = hasil + hasil_convert
    hasil = hasil / 17
    nilai_tiap_kk['CPS'][index] = hasil
    
for index, row in df_wajib_se.iterrows():
    hasil = 0
    for column in df_wajib_se:
        hasil_convert = df_wajib_se.loc[index, column]
        hasil = hasil + hasil_convert
    hasil = hasil / 11
    nilai_tiap_kk['SE'][index] = hasil

for index, row in df_wajib_ds.iterrows():
    hasil = 0
    for column in df_wajib_ds:
        hasil_convert = df_wajib_ds.loc[index, column]
        hasil = hasil + hasil_convert
    hasil = hasil / 14
    nilai_tiap_kk['DS'][index] = hasil

#menyimpan data nilai_tiap_kk ke dalam file total_nilai_tiap_kk.csv
nilai_tiap_kk.to_csv('csv/total_nilai_tiap_kk.csv', index = False)
nilai_tiap_kk

#membuat kolom Kelompok Keahlian pada dataset nilai_tiap_kk
nilai_tiap_kk['KK'] = np.nan

#menentukan Kelompok Keahlian berdasarkan nilai tertinggi pada dataframe nilai_tiap_kk
#dan melakukan penandaan jika terdapat data yang bersifat ambigu
for index, row in nilai_tiap_kk.iterrows():
    if nilai_tiap_kk['IS'][index] > nilai_tiap_kk['CPS'][index] and nilai_tiap_kk['IS'][index] > nilai_tiap_kk['SE'][index] and nilai_tiap_kk['IS'][index] > nilai_tiap_kk['DS'][index]:
        KK = 0
    elif nilai_tiap_kk['CPS'][index] > nilai_tiap_kk['IS'][index] and nilai_tiap_kk['CPS'][index] > nilai_tiap_kk['SE'][index] and nilai_tiap_kk['CPS'][index] > nilai_tiap_kk['DS'][index]:
        KK = 1
    elif nilai_tiap_kk['SE'][index] > nilai_tiap_kk['IS'][index] and nilai_tiap_kk['SE'][index] > nilai_tiap_kk['CPS'][index] and nilai_tiap_kk['SE'][index] > nilai_tiap_kk['DS'][index]:
        KK = 2
    elif nilai_tiap_kk['DS'][index] > nilai_tiap_kk['IS'][index] and nilai_tiap_kk['DS'][index] > nilai_tiap_kk['CPS'][index] and nilai_tiap_kk['DS'][index] > nilai_tiap_kk['SE'][index]:
        KK = 3

    nilai_tiap_kk.loc[index, 'KK'] = KK
    KK = 4 #jika terdapat data yg bersifat ambigu KK = 4

#menggabungkan dataset df_lulus dengan dataset df bagian kolom KK saja
df_lulus = pd.concat([df_lulus, nilai_tiap_kk['KK']], axis=1)
df_lulus

#melakukan pengecekkan banyaknya mahasiswa pada setiap Kelompok Keahlian (IS, CPS, SE, DS)
print('Total Mahasiswa KK IS  = ', len(df_lulus[df_lulus['KK'] == 0]))
print('Total Mahasiswa KK CPS = ', len(df_lulus[df_lulus['KK'] == 1]))
print('Total Mahasiswa KK SE = ', len(df_lulus[df_lulus['KK'] == 2]))
print('Total Mahasiswa KK DS = ', len(df_lulus[df_lulus['KK'] == 3]))

#mengisi nilai null menjadi 4 sebagai penanda bahwa pada data tsb ambigu dan memasukkannya ke dalam dataset ambigu
df_ambigu = pd.DataFrame(columns=df_lulus.columns)
nilai_tiap_kk_ambigu = pd.DataFrame(columns=nilai_tiap_kk.columns) 

#memindahkan data yang bersifat ambigu ke dalam df_ambigu dan melakukan drop data yang bersifat ambigu pada dataset df_lulus
for index, row in nilai_tiap_kk.iterrows():
    if nilai_tiap_kk['KK'][index] == 4:
        df_ambigu = df_ambigu.append(df_lulus.loc[index])
        df_lulus = df_lulus.drop([index])
        
        nilai_tiap_kk_ambigu = nilai_tiap_kk_ambigu.append(nilai_tiap_kk.loc[index])
        
df_ambigu = df_ambigu.drop(['KK'], axis = 1)
df_ambigu = df_ambigu.reset_index(drop=True)
nilai_tiap_kk_ambigu = nilai_tiap_kk_ambigu.reset_index(drop=True)
df_lulus = df_lulus.reset_index(drop=True)
df_lulus

#melakukan pengecekkan banyaknya mahasiswa pada setiap Kelompok Keahlian (IS, CPS, SE, DS) setelah drop data ambigu
print('Total Mahasiswa KK IS  = ', len(df_lulus[df_lulus['KK'] == 0]))
print('Total Mahasiswa KK CPS = ', len(df_lulus[df_lulus['KK'] == 1]))
print('Total Mahasiswa KK SE = ', len(df_lulus[df_lulus['KK'] == 2]))
print('Total Mahasiswa KK DS = ', len(df_lulus[df_lulus['KK'] == 3]))

nilai_tiap_kk_ambigu

#drop data yg memiliki nilai 0 pada setiap prasyarat KK (dianggap sebagai Fake Data)
for index, row in df_lulus.iterrows():
    for column in df_wajib_is:
        df_lulus = df_lulus[df_lulus[column] != 0]
    for column in df_wajib_cps:
        df_lulus = df_lulus[df_lulus[column] != 0]
    for column in df_wajib_se:
        df_lulus = df_lulus[df_lulus[column] != 0]
    for column in df_wajib_ds:
        df_lulus = df_lulus[df_lulus[column] != 0]
df_lulus = df_lulus.reset_index(drop=True)
df_lulus

#menyimpan dataset training dan dataset ambigu
df_lulus.to_csv('csv/data_train1617.csv', index = False)
df_ambigu.to_csv('csv/data_ambigu.csv', index = False)

#melakukan pengecekkan banyaknya mahasiswa pada setiap Kelompok Keahlian (IS, CPS, SE, DS) 
print('Total Mahasiswa KK IS  = ', len(df_lulus[df_lulus['KK'] == 0]))
print('Total Mahasiswa KK CPS = ', len(df_lulus[df_lulus['KK'] == 1]))
print('Total Mahasiswa KK SE = ', len(df_lulus[df_lulus['KK'] == 2]))
print('Total Mahasiswa KK DS = ', len(df_lulus[df_lulus['KK'] == 3]))

#testing
#membuat data training dengan ketentuan 125 data setiap KKnya (undersampling)
IS = 0
CPS = 0
SE = 0
DS = 0

df_new_train = pd.DataFrame(columns=df_lulus.columns)

for index, row in df_lulus.iterrows():
    if df_lulus['KK'][index] == 0 and IS != 125:
        df_new_train = df_new_train.append(df_lulus.loc[index])
        IS = IS + 1
    elif df_lulus['KK'][index] == 1 and CPS != 125:
        df_new_train = df_new_train.append(df_lulus.loc[index])
        CPS = CPS + 1
    elif df_lulus['KK'][index] == 2 and SE != 125:
        df_new_train = df_new_train.append(df_lulus.loc[index])
        SE = SE + 1
    elif df_lulus['KK'][index] == 3 and DS != 125:
        df_new_train = df_new_train.append(df_lulus.loc[index])
        DS = DS + 1
        
df_new_train = df_new_train.reset_index(drop=True)

df_new_train.to_csv('csv/data_train1617_undersampling.csv', index = False)
df_new_train

import pandas_profiling
from pandas_profiling import ProfileReport

#eksplorasi Data Menggunakan Pandas Profiling Report

pandas_report = df_lulus.profile_report(sort='None', html={'style':{'full_width': True}})
pandas_report
